% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funcs.R
\name{DGP}
\alias{DGP}
\title{Data generate process}
\usage{
DGP(Model, case, n = 200)
}
\arguments{
\item{Model}{choose model 1: linear model, 2: poisson model}

\item{case}{choose y case 1-4}

\item{n}{sample size, default = 200}
}
\value{
a list contains X and Y.
}
\description{
Data generate process
}
\examples{
data = DGP(Model = 1, case = 1)
x = data$X
y = data$Y
}
