% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funcs.R
\name{para.solve}
\alias{para.solve}
\title{Estimate model parameters}
\usage{
para.solve(x, y, Model)
}
\arguments{
\item{x}{data X, n*1 vector}

\item{y}{data Y, n*1 vector}

\item{Model}{choose model 1: linear model, 2: poisson model}
}
\value{
the estimation of unknown parameters.
}
\description{
Estimate model parameters
}
\examples{
data = DGP(Model = 1, case = 1)
x = data$X
y = data$Y
theta.hat = para.solve(x, y, Model = 1)
}
