% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funcs.R
\name{Tn.b}
\alias{Tn.b}
\title{Bootstrap test statistic sample}
\usage{
Tn.b(x, y.b, c.vec, Model, B, Phi.func = NULL)
}
\arguments{
\item{x}{data X, n*1 vector}

\item{y.b}{bootstrap sample, B*1 list}

\item{c.vec}{the vector contains all parameter c}

\item{Model}{choose model 1: linear model, 2: poisson model}

\item{B}{the number of bootstrap}

\item{Phi.func}{the one-to-one mapping Phi}
}
\value{
a matrix contains: bootstrap test statistics in different c and max value.
}
\description{
Bootstrap test statistic sample
}
\examples{
data = DGP(Model = 1, case = 1)
x = data$X
y = data$Y
theta.hat = para.solve(x, y, Model = 1)
y.tilde = y.tilde.sample(x, theta.hat, Model = 1)
Tn.hat = Tnsc(x, y, y.tilde, c = 5)
y.b = bootstrap.sample(x, theta.hat, Model = 1, B = 500)
Tn.b.hat = Tn.b(x, y.b, c.vec = 5, Model = 1, B = 500, Phi.func = NULL)
}
