% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funcs.R
\name{y.tilde.sample}
\alias{y.tilde.sample}
\title{Sampling y.tilde}
\usage{
y.tilde.sample(x, theta.hat, Model)
}
\arguments{
\item{x}{data X, n*1 vector}

\item{theta.hat}{the QMLE}

\item{Model}{choose model 1: linear model, 2: poisson model}
}
\value{
y.tilde, n*1 vector.
}
\description{
Sampling y.tilde
}
\examples{
data = DGP(Model = 1, case = 1)
x = data$X
y = data$Y
theta.hat = para.solve(x, y, Model = 1)
y.tilde = y.tilde.sample(x, theta.hat, Model = 1)
}
